web handlers moved to graphics subsystem
