+++ Source/Urho3D/Graphics/OpenGL/OGLGraphics.cpp
@@ -53,6 +53,10 @@
 #endif
 
 #ifdef __EMSCRIPTEN__
+#include "../UI/UI.h"
+#include <emscripten/emscripten.h>
+#include <emscripten/bind.h>
+
 // Emscripten provides even all GL extension functions via static linking. However there is
 // no GLES2-specific extension header at the moment to include instanced rendering declarations,
 // so declare them manually from GLES3 gl2ext.h. Emscripten will provide these when linking final output.
@@ -61,6 +65,25 @@ extern "C"
     GL_APICALL void GL_APIENTRY glDrawArraysInstancedANGLE (GLenum mode, GLint first, GLsizei count, GLsizei primcount);
     GL_APICALL void GL_APIENTRY glDrawElementsInstancedANGLE (GLenum mode, GLsizei count, GLenum type, const void *indices, GLsizei primcount);
     GL_APICALL void GL_APIENTRY glVertexAttribDivisorANGLE (GLuint index, GLuint divisor);
+}
+
+// Helper functions to support emscripten canvas resolution change
+static const Urho3D::Context *appContext;
+
+static void JSCanvasSize(int width, int height, bool fullscreen, float scale)
+{
+    URHO3D_LOGINFOF("JSCanvasSize: width=%d height=%d fullscreen=%d scale=%f", width, height, fullscreen, scale);
+    using namespace Urho3D;
+    if (appContext)
+        appContext->GetSubsystem<Graphics>()->SetMode(width, height);
+    UI* ui = appContext->GetSubsystem<UI>();
+    if (ui)
+        ui->SetScale(scale);
+}
+
+using namespace emscripten;
+EMSCRIPTEN_BINDINGS(Module) {
+    function("JSCanvasSize", &JSCanvasSize);
 }
 #endif
 
@@ -242,6 +265,10 @@ Graphics::Graphics(Context* context) :
 
     // Register Graphics library object factories
     RegisterGraphicsLibrary(context_);
+
+#ifdef __EMSCRIPTEN__
+        appContext = context_;
+#endif
 }
 
 Graphics::~Graphics()
@@ -340,8 +367,11 @@ bool Graphics::SetScreenMode(int width, int height, const ScreenModeParams& para
             flags |= SDL_WINDOW_BORDERLESS;
         if (newParams.resizable_)
             flags |= SDL_WINDOW_RESIZABLE;
+
+#ifndef __EMSCRIPTEN__
         if (newParams.highDPI_)
             flags |= SDL_WINDOW_ALLOW_HIGHDPI;
+#endif
 
         SDL_SetHint(SDL_HINT_ORIENTATIONS, orientations_.CString());
 
@@ -2151,6 +2181,12 @@ void Graphics::OnWindowResized()
     ResetRenderTargets();
 
     URHO3D_LOGDEBUGF("Window was resized to %dx%d", width_, height_);
+
+#ifdef __EMSCRIPTEN__
+        EM_ASM({
+            Module.SetRendererSize($0, $1);
+        }, width_, height_);
+#endif
 
     using namespace ScreenMode;
